pyproject.toml
[project]
name = "ixora-meeting-booking"
version = "0.1.0"
description = "AI agent for booking meetings with Ixora Solution's CEO via Microsoft Bookings"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.1.0",
    "langgraph>=0.0.40",
    "langchain-google-genai>=1.0.0",
    "playwright>=1.40.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]
